refactor: :truck: move interfaces to own folder
