+++ src/interfaces/Header.interface.ts
@@ -1,0 +1,7 @@
+export interface IHeader {}
+
+export interface IHeaderItems {
+  url: string
+  label: string
+  id: string
+}
