+++ src/interfaces/Icon.interface.ts
@@ -1,0 +1,7 @@
+export interface IIconProps {
+  icon: string
+  className?: string
+  width?: string
+  height?: string
+  onClick?: () => void
+}
