+++ src/interfaces/Layout.interface.ts
@@ -1,0 +1,3 @@
+export interface ILayout {
+  path: string
+}
